refactor: rename Types folder and files to lowercase
