--- src/components/Types/TodoAggregate.ts
@@ -1,7 +1,0 @@
-import { Nullable } from './Nullable';
-import { Todo } from './Todo';
-import { User } from './User';
-
-export type TodoAggregate = Todo & {
-  user: Nullable<User>;
-};
